Add Husky pre-commit hook and update Cypress configuration***
This commit adds a new Husky pre-commit hook that runs the "npm test" command before committing changes. It also updates the Cypress configuration file to include a new "baseUrl" property set to "http://localhost:3000". Additionally, the navigation test in the "navigation.cy.ts" file is updated to include an assertion that checks for the presence of an "h1" element containing the text "Joshua Hoban". The package-lock.json and package.json files are also updated to include the latest version of Husky as a dev dependency. Finally, the tsconfig.json file is updated to include the "cypress" type in the "types" array.